@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <UserSecretsId>4761b4d0-7f3f-4a3a-b774-89de95b85112</UserSecretsId>
   </PropertyGroup>
 
@@ -14,10 +14,12 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App"  Version="2.1.4" AllowExplicitVersion="true" />
-    <PackageReference Include="Ocelot" Version="12.0.1" />
-    <PackageReference Include="Ocelot.Cache.CacheManager" Version="0.1.0" />
-    <PackageReference Include="Ocelot.Provider.Eureka" Version="0.1.2" />
+    <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="5.5.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="3.1.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.0" />
+    <PackageReference Include="Ocelot" Version="13.8.0" />
+    <PackageReference Include="Ocelot.Cache.CacheManager" Version="13.8.0" />
+    <PackageReference Include="Ocelot.Provider.Eureka" Version="13.8.0" />
   </ItemGroup>
 
 </Project>
