@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using ProductService.DataAccess.EF;
 using ProductService.Init;
 using Steeltoe.Discovery.Client;
@@ -25,15 +26,17 @@ namespace ProductService
             services.AddDiscoveryClient(Configuration);
             services.AddEFConfiguration(Configuration);
             services.AddMvc()
-                .AddJsonOptions(JsonOptions)
-                .SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+                .AddNewtonsoftJson(JsonOptions)
+                .SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
             services.AddMediatR();
             services.AddProductDemoInitializer();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
+            app.UseRouting();
+            
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
@@ -44,12 +47,12 @@ namespace ProductService
             }
 
             app.UseHttpsRedirection();
-            app.UseMvc();
             app.UseInitializer();
             app.UseDiscoveryClient();
+            app.UseEndpoints(endpoints => { endpoints.MapControllers(); });
         }
         
-        private void JsonOptions(MvcJsonOptions options)
+        private void JsonOptions(MvcNewtonsoftJsonOptions options)
         {
             options.SerializerSettings.Converters.Add(new Newtonsoft.Json.Converters.StringEnumConverter());
             options.SerializerSettings.NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore;
