@@ -1,28 +1,21 @@
-﻿using MediatR;
-using PolicyService.Api.Commands;
-using PolicyService.Api.Commands.Dtos;
-using PolicyService.Domain;
-using PolicyService.RestClients;
-using PricingService.Api.Commands;
-using PricingDtos = PricingService.Api.Commands.Dto;
-using RestEase;
-using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using NHibernate;
+using MediatR;
+using PolicyService.Api.Commands;
+using PolicyService.Domain;
+using PricingDtos = PricingService.Api.Commands.Dto;
 
 namespace PolicyService.Commands
 {
     public class CreateOfferHandler : IRequestHandler<CreateOfferCommand, CreateOfferResult>
     {
-        private readonly IUnitOfWorkProvider uowProvider;
+        private readonly IUnitOfWork uow;
         private readonly IPricingService pricingService;
 
-        public CreateOfferHandler(IUnitOfWorkProvider uowProvider, IPricingService pricingService)
+        public CreateOfferHandler(IUnitOfWork uow, IPricingService pricingService)
         {
-            this.uowProvider = uowProvider;
+            this.uow = uow;
             this.pricingService = pricingService;
         }
 
@@ -42,14 +35,11 @@ namespace PolicyService.Commands
                 );
 
             //create and save offer
-            using (var uow = uowProvider.Create())
-            {
-                uow.Offers.Add(o);
-                await uow.CommitChanges();
+            uow.Offers.Add(o);
+            await uow.CommitChanges();
 
-                //return result
-                return ConstructResult(o);
-            }
+            //return result
+            return ConstructResult(o);
         }
 
         private CreateOfferResult ConstructResult(Offer o)
