@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading;
 using System.Threading.Tasks;
 
 namespace PolicyService.Domain
@@ -14,9 +11,5 @@ namespace PolicyService.Domain
 
         Task CommitChanges();
     }
-
-    public interface IUnitOfWorkProvider
-    {
-        IUnitOfWork Create();
-    }
+    
 }
